Group renewable electricity types in stacked bar chart
Previously we always ordered the labeled subcategories in the stacked bar chart alphabetically, but that meant that renewable energy and coal and oil got mixed. It's easier to read when all polluting energy types are together, and all renewable types are together.

I decided to allow the data file to specify the order, so no one has to modify this code to change the orders in the future.
The order only applies to the "sector" breakdown now, because otherwise it would reorder the "region" breakdown to be out of alphabetical order, which would look weird.