Prepare to read multiple files as different scenarios
We want each file in the data folder to display as a separate scenario, titled by the name of the file it was loaded from. Previously, the code was set up to read all scenarios from one file. This prepares it to load each scenario from a separate file. The next commit will finish that work.