Filter out missing subcategories from subcategory dropdown
Do not show a subcategory in the dropdown if all the values for it are NULL.
It would be nice to be able to do this filtering without a for loop, but I'll
leave that as a future refactor. For now it's working.

Fixes https://github.com/cypressf/mit-climate-data-viz/issues/25